feat: #WB2-1511, add AppActions based on rights
