+++ frontend/src/store/index.ts
@@ -1,2 +1,3 @@
 export * from './rights';
 export * from './treeview';
+export * from './wiki';
